app/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
pydantic-settings 
SQLAlchemy        
psycopg2-binary    
alembic
python-jose